# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace clothing brand e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Variant {
  id: string;
  productId: string;
  size: string;
  color: string;
  price: number;
  stock: number;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  basePrice: number;
  isActive: boolean;
  images: string[];
  variants: Variant[];
}

export interface OrderItemInput {
  variantId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface CreateOrderInput {
  /** @minLength 3 */
  customerName: string;
  /** @minLength 6 */
  phone: string;
  /** @minLength 6 */
  paymentPhone: string;
  /** @minLength 10 */
  address: string;
  /** @minLength 2 */
  city: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface OrderItem {
  id: string;
  orderId: string;
  productName: string;
  size: string;
  color: string;
  quantity: number;
  price: number;
}

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  Pending: "Pending",
  Paid: "Paid",
} as const;

export type OrderOrderStatus =
  (typeof OrderOrderStatus)[keyof typeof OrderOrderStatus];

export const OrderOrderStatus = {
  Pending: "Pending",
  Shipped: "Shipped",
  Delivered: "Delivered",
  Cancelled: "Cancelled",
} as const;

export interface Order {
  id: string;
  customerName: string;
  phone: string;
  paymentPhone: string;
  address: string;
  city: string;
  totalAmount: number;
  paymentStatus: OrderPaymentStatus;
  orderStatus: OrderOrderStatus;
  createdAt: string;
  items: OrderItem[];
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  NewOrder: "NewOrder",
  PaymentReceived: "PaymentReceived",
} as const;

export interface Notification {
  id: string;
  type: NotificationType;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface AdminStats {
  total: number;
  pending: number;
  paid: number;
  shipped: number;
  delivered: number;
  unreadNotifications: number;
}

export type AdminLoginBody = {
  username: string;
  password: string;
};

export type AdminLogin200 = {
  token: string;
};

export type ListOrdersParams = {
  token: string;
};

export type GetOrderParams = {
  token: string;
};

export type ConfirmPaymentBody = {
  token: string;
};

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  Pending: "Pending",
  Shipped: "Shipped",
  Delivered: "Delivered",
  Cancelled: "Cancelled",
} as const;

export type UpdateOrderStatusBody = {
  status: UpdateOrderStatusBodyStatus;
  token: string;
};

export type ListNotificationsParams = {
  token: string;
};

export type MarkNotificationReadBody = {
  token: string;
};

export type GetAdminStatsParams = {
  token: string;
};