# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace clothing brand e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  total: number;
  pending: number;
  paid: number;
  shipped: number;
  delivered: number;
  unreadNotifications: number;
}