# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace clothing brand e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderOrderStatus } from "./orderOrderStatus";
import type { OrderPaymentStatus } from "./orderPaymentStatus";

export interface Order {
  id: string;
  customerName: string;
  phone: string;
  paymentPhone: string;
  address: string;
  city: string;
  totalAmount: number;
  paymentStatus: OrderPaymentStatus;
  orderStatus: OrderOrderStatus;
  createdAt: Date;
  items: OrderItem[];
}