# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace clothing brand e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: string;
  orderId: string;
  productName: string;
  size: string;
  color: string;
  quantity: number;
  price: number;
}