# lib/api-zod/src/generated/types/variant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Peace clothing brand e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Variant {
  id: string;
  productId: string;
  size: string;
  color: string;
  price: number;
  stock: number;
}